+++ authority-portal-backend/authority-portal-quarkus/src/main/kotlin/de/sovity/authorityportal/web/pages/usermanagement/UserInvitationApiService.kt
@@ -3,7 +3,6 @@ package de.sovity.authorityportal.web.pages.usermanagement
 import de.sovity.authorityportal.api.model.IdResponse
 import de.sovity.authorityportal.api.model.InviteParticipantUserRequest
 import de.sovity.authorityportal.db.jooq.enums.UserRegistrationStatus
-import de.sovity.authorityportal.web.pages.usermanagement.UserRoleMapper
 import de.sovity.authorityportal.web.services.UserService
 import de.sovity.authorityportal.web.thirdparty.keycloak.KeycloakService
 import io.quarkus.logging.Log