+++ authority-portal-backend/authority-portal-quarkus/src/main/kotlin/de/sovity/authorityportal/web/services/OrganizationService.kt
@@ -4,6 +4,7 @@ import de.sovity.authorityportal.api.model.CreateOrganizationRequest
 import de.sovity.authorityportal.db.jooq.Tables
 import de.sovity.authorityportal.db.jooq.enums.OrganizationRegistrationStatus
 import de.sovity.authorityportal.db.jooq.tables.records.OrganizationRecord
+import de.sovity.authorityportal.web.model.CreateOrganizationData
 import jakarta.enterprise.context.ApplicationScoped
 import jakarta.inject.Inject
 import org.jooq.DSLContext
@@ -44,9 +45,39 @@ class OrganizationService {
             it.mdsId = mdsId
             it.name = organization.name
             it.address = organization.address
-            it.duns = organization.duns
+            it.taxId = organization.taxId
             it.url = organization.url
-            it.securityEmail = organization.securityEmail
+            it.mainContactEmail = organization.securityEmail
+            it.createdBy = userId
+            it.registrationStatus = registrationStatus
+            it.createdAt = OffsetDateTime.now()
+
+            it.insert()
+        }
+    }
+
+    fun createOrganization(
+        userId: String,
+        mdsId: String,
+        organizationData: CreateOrganizationData,
+        registrationStatus: OrganizationRegistrationStatus
+    ) {
+        dsl.newRecord(Tables.ORGANIZATION).also {
+            it.mdsId = mdsId
+            it.name = organizationData.name
+            it.url = organizationData.url
+            it.businessUnit = organizationData.businessUnit
+            it.address = organizationData.address
+            it.billingAddress = organizationData.billingAddress
+            it.taxId = organizationData.taxId
+            it.commerceRegisterNumber = organizationData.commerceRegisterNumber
+            it.commerceRegisterLocation = organizationData.commerceRegisterLocation
+            it.mainContactName = organizationData.mainContactName
+            it.mainContactEmail = organizationData.mainContactEmail
+            it.mainContactPhone = organizationData.mainContactPhone
+            it.techContactName = organizationData.techContactName
+            it.techContactEmail = organizationData.techContactEmail
+            it.techContactPhone = organizationData.techContactPhone
             it.createdBy = userId
             it.registrationStatus = registrationStatus
             it.createdAt = OffsetDateTime.now()