+++ authority-portal-backend/authority-portal-quarkus/src/test/kotlin/de/sovity/authorityportal/web/integration/pages/TestData.kt
@@ -8,7 +8,7 @@ object TestData {
     const val USER_PHONE_NUMBER = "+1234567890"
     const val ORG_NAME = "Organization ABC"
     const val ORG_ADDRESS = "123 Main St, City, Country"
-    const val ORG_DUNS = "123456789"
+    const val ORG_TAX_ID = "123456789"
     const val ORG_URL = "https://organization-abc.com"
     const val ORG_SECURITY_EMAIL = "security@organization-abc.com"
 }